move load_config into config.py
